pyproject.toml
[project]
name = "reddit-scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=5.5",
    "praw>=7.8.1",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
